Provide noop API for web environments
