+++ src/NativeFullStory.ts
@@ -1,6 +1,6 @@
 import type { TurboModule } from 'react-native';
 import { TurboModuleRegistry } from 'react-native';
-import { OnReadyResponse } from '.';
+import { OnReadyResponse } from './fullstoryInterface';
 
 export interface Spec extends TurboModule {
   anonymize(): void;
