+++ src/index.ts
@@ -3,6 +3,7 @@ import codegenNativeCommands from 'react-native/Libraries/Utilities/codegenNativ
 import type { ViewProps } from 'react-native/Libraries/Components/View/ViewPropTypes';
 import { ForwardedRef } from 'react';
 import { isTurboModuleEnabled } from './utils';
+import { FullstoryStatic, LogLevel } from './fullstoryInterface';
 
 interface NativeProps extends ViewProps {
   fsClass?: string;
@@ -36,43 +37,6 @@ const FullStoryPrivate = isTurboModuleEnabled
   ? require('./NativeFullStoryPrivate').default
   : NativeModules.FullStoryPrivate;
 
-export enum LogLevel {
-  Log = 0, // Clamps to Debug on iOS
-  Debug = 1,
-  Info = 2, // Default
-  Warn = 3,
-  Error = 4,
-  Assert = 5, // Clamps to Error on Android
-}
-
-interface UserVars {
-  displayName?: string;
-  email?: string;
-  [key: string]: any;
-}
-
-export type OnReadyResponse = {
-  replayStartUrl: string;
-  replayNowUrl: string;
-  sessionId: string;
-};
-
-declare type FullStoryStatic = {
-  LogLevel: typeof LogLevel;
-  anonymize(): void;
-  identify(uid: string, userVars?: UserVars): void;
-  setUserVars(userVars: UserVars): void;
-  onReady(): Promise<OnReadyResponse>;
-  getCurrentSession(): Promise<string>;
-  getCurrentSessionURL(): Promise<string>;
-  consent(userConsents: boolean): void;
-  event(eventName: string, eventProperties: Object): void;
-  shutdown(): void;
-  restart(): void;
-  log(logLevel: LogLevel, message: string): void;
-  resetIdleTimer(): void;
-};
-
 declare type FullStoryPrivateStatic = {
   onFSPressForward?(
     tag: number,
@@ -81,16 +45,6 @@ declare type FullStoryPrivateStatic = {
     hasLongPressHandler: boolean,
   ): void;
 };
-
-declare global {
-  namespace JSX {
-    interface IntrinsicAttributes {
-      fsAttribute?: { [key: string]: string };
-      fsClass?: string;
-      fsTagName?: string;
-    }
-  }
-}
 
 const identifyWithProperties = (uid: string, userVars = {}) => identify(uid, userVars);
 
@@ -187,7 +141,7 @@ export function applyFSPropertiesWithRef(existingRef?: ForwardedRef<unknown>) {
   };
 }
 
-const FullStoryAPI: FullStoryStatic = {
+const FullstoryAPI: FullstoryStatic = {
   anonymize,
   identify: identifyWithProperties,
   setUserVars,
@@ -206,4 +160,4 @@ const FullStoryAPI: FullStoryStatic = {
 export const PrivateInterface: FullStoryPrivateStatic =
   Platform.OS === 'android' ? { onFSPressForward: FullStoryPrivate.onFSPressForward } : {};
 
-export default FullStoryAPI;
+export default FullstoryAPI;
